pyproject.toml
[tool.poetry]
name = "cellforge"
version = "0.1.0"
description = ""
authors = ["tsak1998 <tsakalis.pantelis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
warp-lang = {extras = ["extras"], version = "^1.3.3"}
notebook = "^7.2.2"
pyvista = "^0.44.1"
vtk = "^9.3.1"
tetgen = "^0.6.5"
trame = "^3.6.5"
pydantic = "^2.9.2"
scipy = "^1.14.1"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
openpyxl = "^3.1.5"
seaborn = "^0.13.2"
scikit-learn = "^1.5.2"
copula = "^0.0.4"
copulas = "^0.11.1"
statsmodels = "^0.14.3"
pyyaml = "^6.0.2"
tqdm = "^4.66.5"
taichi = "^1.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
